fix logic ImportOrderDetail

+++ product-service/src/main/java/com/unicoin/product/form/AddImportOrderDetail.java
@@ -10,9 +10,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AddImportOrderDetail {
-    private Integer varianId;
+    private Long varianId;
 
     private Integer quantity;
 
-    private Integer cost;
+    private Long cost;
 }
